cnm: Add Chips&Media wave633c firmware for NXP i.MX9
Add the firmware for the Chips&Media wave633c codec IP,
which integrates both decoder and encoder functionality.
This firmware is intended for use on NXP SoCs based on
the i.MX9 architecture.

The LICENSE.cnm file has been updated to restrict redistribution
of this firmware to NXP silicon products only.

FW version : 4.0.0

Signed-off-by: Nas Chung <nas.chung@chipsnmedia.com>