Add whence2yaml.py and helper library
Add a script and helper library that convert the WHENCE file to YAML
format and write the data to WHENCE.yaml. This can we be used as a
starting point for making the WHENCE data better machine-parseable and
also serves as a foundation for adding more relavant data for firmware
consumers like distros.

Signed-off-by: Juerg Haefliger <juerg.haefliger@canonical.com>